feat: progress on index and partitions
