+++ docker/servers_local.json
@@ -3,7 +3,7 @@
         "1": {
             "Group": "Servers",
             "Name": "Docker",
-            "Host": "postgres",
+            "Host": "localhost",
             "Port": 5432,
             "MaintenanceDB": "postgres",
             "Username": "postgres",
