+++ exec/configs.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+export schema_definitions_path="schemas/definitions/sinch_db/"
+export db_config_path="docker/servers_local.json"
