+++ ingestion/schemas/definitions/sinch_db/order_items.yml
@@ -1,0 +1,18 @@
+database: postgres
+schema: public
+table: orders
+columns:
+  order_id:
+    type: "VARCHAR(100)"
+    constraints: ["NOT NULL"]
+  item_name:
+    type: "VARCHAR(255)"
+  price:
+    type: "DECIMAL(10,2)"
+indexes:
+  - name: "idx_order_id"
+    columns: ["order_id"]
+  - name: "idx_item_name"
+    columns: ["item_name"]
+  - name: "idx_price"
+    columns: ["price"]
